Incorporates changes to crawler.py and moves tests to test/ directory
